Sensor value (#82)
* Add buffer for SensorReading class

* SFR Conversion to SFRField<>

* updated SensorReading methods

* Remove remaining SensorReadings from SFR

* Removed sfr.cpp

* Moved variables from SFR into classes

* Add imu downlink to class

* Moved sensor reading values for IMUMonitor

* Remove imu dependencies and modify IMUDownlinkReportMonitor to use sfr

* Formatting and moved imu report written boolean

* Store current mission modes in sfr

* Updated MissionManager. Make sensor type an enum class.

* Move sensor values from battery, current, photoresistor, and temperature monitors

* Formatting

* Update mission modes with new sensor reading methods

* FS-146: Initial Rockblock and camera changes

* resolve remaining camera issues

* included pins.hpp and handled enum cast to int

* Fix multiple definitions issue with SFR

fix formatting

* handle eeprom string bug

* Pull request feedback.

* Formatting

* Formatting

* Formatting

+++ src/Control Tasks/BurnwireControlTask.cpp
@@ -93,7 +93,7 @@ void BurnwireControlTask::execute()
 void BurnwireControlTask::dispatch_burn()
 {
     sfr::burnwire::camera_attempts = 0;
-    sfr::burnwire::mode = burnwire_mode_type::burn;
+    sfr::burnwire::mode = (uint16_t)burnwire_mode_type::burn;
     sfr::burnwire::start_time = millis();
     if (sfr::burnwire::attempts > constants::burnwire::max_attempts) {
         transition_to_standby();
@@ -111,7 +111,7 @@ void BurnwireControlTask::dispatch_burn()
 
 void BurnwireControlTask::transition_to_standby()
 {
-    sfr::burnwire::mode = burnwire_mode_type::standby;
+    sfr::burnwire::mode = (uint16_t)burnwire_mode_type::standby;
     Pins::setPinState(constants::burnwire::first_pin, LOW);
     Pins::setPinState(constants::burnwire::second_pin, LOW);
     sfr::burnwire::fire = false;
